+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * 0.1.9
     * reorder system logs fields for charts
-    * ssh versions checks
+    * ssh versions checks (for ubuntu)
+    * system support check (for ubuntu)
 
 * 0.1.8
     * fix ssh logs parsing